pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llmsec"
version = "0.1.0"
description = "A Python library for LLM security"
authors = [
    { name="Greg Retkowski", email="greg@rage.net" }
]
urls = { Homepage = "https://github.com/gregretkowski/llmsec" }
dependencies = [
    "litellm",
    "python-dotenv",
    "jinja2"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9"
]


[project.scripts]
check-prompt = "llmsec:cli"
[project.optional-dependencies]
dev = [
    "pylint",
    "pytest",
    "httpretty"
]
